requirements.txt
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
torch==2.0.1
navec==0.10.0
joblib==1.3.2
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
python-multipart==0.0.6
pyyaml==6.0
joblib==1.3.2